road_model/requirements.txt
grpcio>=1.47.0
protobuf>=1.20
numpy>=1.19
geodesic-api>=0.6.0
torch==1.12.0